requirements.txt
playwright>=1.49.0
pandas>=2.2.3
python-dotenv==1.0.1
beautifulsoup4==4.12.3
jinja2==3.1.3
